% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_roxygen_params_from_file.R
\name{indent_from_param}
\alias{indent_from_param}
\title{helper function to indent secondary lines of text, based on nchar of param}
\usage{
indent_from_param(param)
}
\arguments{
\item{param}{the name of a function parameter whose description text will be indented in the documentation}
}
\description{
helper function to indent secondary lines of text, based on nchar of param
}
