% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_roxygen_params_from_file.R
\name{make_param_docs_from_file}
\alias{make_param_docs_from_file}
\title{Takes a file and creates a single, multi-paragraph string of all the params in an appropriate syntax}
\usage{
make_param_docs_from_file(file_full_path, okay_to_source.safety_flag = FALSE)
}
\arguments{
\item{file_full_path}{The file which will be documented}

\item{okay_to_source.safety_flag}{A TRUE/FALSE flag. Setting to FALSE will cause function to terminate in error without reading in the file.}
}
\description{
Takes a file and creates a single, multi-paragraph string of all the params in an appropriate syntax
}
