% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen_documentation_template_and_format.R
\name{roxygen_documentation_template_and_format}
\alias{roxygen_documentation_template_and_format}
\alias{catn}
\alias{make_group_documentation}
\alias{add_roxygen_ticks}
\alias{neat_box}
\title{Roxygen Documentation Template and Format}
\usage{
catn(..., sep = "\\n")

make_group_documentation(file_full_path,
  one_liner = "ONE-LINER WHAT DO THESE GROUP OF FUNCS DO? (or the name of the main function)",
  detailed_desc = "DETAILED DESCRIPTION of what these functions do",
  other_sections = list(section_name = "text paragraph"), aliases = c(),
  export_func = "..auto..", return = "WHAT IS RETURNED??",
  example_text = "\\n", filename = basename(file_full_path),
  title = filename, header = filename, okay_to_source.safety_flag = FALSE)

add_roxygen_ticks(x, tick = "#' ", clear_multiple_lines = FALSE,
  at_least_reps = 2L, comment_space = "X")

neat_box(x, min_width = 62L, left_pad = 3L, collapse = "\\n")
}
\arguments{
\item{...}{<DESCRIBE ME>
<WHAT IS IT?>}

\item{sep}{<DESCRIBE ME>
           <WHAT IS IT?>

           Defaults to: "\n"}

\item{file_full_path}{<DESCRIBE ME>
<WHAT IS IT?>}

\item{one_liner}{<DESCRIBE ME>
                 <WHAT IS IT?>

                 Defaults to: "ONE-LINER WHAT DO THESE GROUP OF FUNCS DO? (or the name of the main function)"}

\item{detailed_desc}{<DESCRIBE ME>
                     <WHAT IS IT?>

                     Defaults to: "DETAILED DESCRIPTION of what these functions do"}

\item{other_sections}{<DESCRIBE ME>
                      <WHAT IS IT?>

                      Defaults to: list(section_name = \"text paragraph\")}

\item{aliases}{<DESCRIBE ME>
               <WHAT IS IT?>

               Defaults to: c()}

\item{export_func}{<DESCRIBE ME>
                   <WHAT IS IT?>

                   Defaults to: "..auto.."}

\item{return}{<DESCRIBE ME>
              <WHAT IS IT?>

              Defaults to: "WHAT IS RETURNED??"}

\item{example_text}{<DESCRIBE ME>
                    <WHAT IS IT?>

                    Defaults to: "\n"}

\item{filename}{<DESCRIBE ME>
                <WHAT IS IT?>

                Defaults to: basename(file_full_path)}

\item{title}{<DESCRIBE ME>
             <WHAT IS IT?>

             Defaults to: filename}

\item{header}{<DESCRIBE ME>
              <WHAT IS IT?>

              Defaults to: filename}

\item{okay_to_source.safety_flag}{<DESCRIBE ME>
                                  <WHAT IS IT?>

                                  Defaults to: FALSE}

\item{x}{<DESCRIBE ME>
<WHAT IS IT?>}

\item{tick}{<DESCRIBE ME>
            <WHAT IS IT?>

            Defaults to: "#' "}

\item{clear_multiple_lines}{<DESCRIBE ME>
                            <WHAT IS IT?>

                            Defaults to: FALSE}

\item{at_least_reps}{<DESCRIBE ME>
                     <WHAT IS IT?>

                     Defaults to: 2L}

\item{min_width}{<DESCRIBE ME>
                 <WHAT IS IT?>

                 Defaults to: 62L}

\item{left_pad}{<DESCRIBE ME>
                <WHAT IS IT?>

                Defaults to: 3L}

\item{collapse}{<DESCRIBE ME>
                <WHAT IS IT?>

                Defaults to: "\n"}
}
\value{
WHAT IS RETURNED??
}
\description{
ONE-LINER WHAT DO THESE GROUP OF FUNCS DO? (or the name of the main function)
}
\details{
DETAILED DESCRIPTION of what these functions do
}
\examples{


}
